+++ lib/recipes/web-generic/recipe.rb
@@ -15,6 +15,7 @@ WEB_RECIPE = Kitchen::BookRecipe.new(book_short_name: :web) do |doc, resources|
   book_pages.each do |page_or_composite|
     BakeOrderHeaders.v1(within: page_or_composite)
   end
+  WebIndexCleanup.v1(book_pages: book_pages)
 
   # Restore
   WebPostBakeRestore.v1(book_pages: book_pages)